feat(openai): migrate Realtime provider from Beta to GA API with gpt-realtime-1.5
OpenAI is removing the Realtime API Beta on May 7, 2026. This migrates
to the GA API: removes the OpenAI-Beta header, adds required session
type field, updates default model to gpt-realtime-1.5, drops the
unsupported temperature parameter, and pins openai SDK to >=2.28.

+++ src/roomkit/video/base.py
@@ -69,7 +69,7 @@ class VideoChunk:
     codec: str = "h264"
     width: int = 640
     height: int = 480
-    timestamp_ms: int | None = None
+    timestamp_ms: float | None = None
     keyframe: bool = False
     is_final: bool = False
